+++ dev/boot/src/boot.ts
@@ -16,25 +16,19 @@
 import { Builder } from '@anticrm/model'
 
 import { model as modelModel } from '@anticrm/model/src/__model__'
-import { model as core } from '@anticrm/core/src/__model__'
 import { model as presentation } from '@anticrm/presentation/src/__model__'
 import { model as contact } from '@anticrm/contact/src/__model__'
 import { model as workbench } from '@anticrm/workbench/src/__model__'
 import { model as task } from '@anticrm/task/src/__model__'
 import { model as chunter } from '@anticrm/chunter/src/__model__'
-// import recruitmentModel from '@anticrm/recruitment-model/src/model'
-
-// import taskStrings from '@anticrm/task-model/src/strings/ru'
 
 export const builder = new Builder()
 builder.load(modelModel)
-builder.load(core)
 builder.load(presentation)
 builder.load(contact)
 builder.load(workbench)
 builder.load(chunter)
 builder.load(task)
-// builder.load(recruitmentModel)
 
 export const Model = builder.dumpAll()
 export const Strings = {
