+++ dev/boot/src/index.ts
@@ -15,7 +15,7 @@
 
 import { MongoClient } from 'mongodb'
 import { Model, Strings } from './boot'
-import { Doc } from '@anticrm/model'
+import { Doc } from '@anticrm/core'
 
 const modelJson = JSON.stringify(Model, null, 2)
 console.log(modelJson)
