+++ plugins/workbench/src/components/internal/Spotlight.svelte
@@ -1,14 +1,14 @@
 <!--
 // Copyright © 2020 Anticrm Platform Contributors.
-// 
+//
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
 // obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// 
+//
 // See the License for the specific language governing permissions and
 // limitations under the License.
 -->
@@ -22,11 +22,12 @@
   import { find } from '../../utils'
 
   import Icon from '@anticrm/platform-ui/src/components/Icon.svelte'
+  import { CORE_CLASS_TITLE } from '@anticrm/domains'
 
   let query: string
   let result: Title[] = []
 
-  $: find(core.class.Title, { title: query as Property<string, string> }).then(docs => { console.log('search', docs); result = docs })
+  $: find(CORE_CLASS_TITLE, { title: query as Property<string, string> }).then(docs => { console.log('search', docs); result = docs })
 </script>
 
 <Icon icon={ui.icon.Search} clazz="icon-embed-2x" />&nbsp;
